move create config script to submission
